Move filterAndGroupCostOverridesForDisplay to wpcom-checkout
